requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
exceptiongroup==1.3.1
fastapi==0.127.0
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
pydantic-settings==2.12.0 
python-dotenv==1.2.1
click==8.3.1
colorama==0.4.6
h11==0.16.0
uvicorn==0.40.0
certifi==2025.11.12
httpcore==1.0.9
httpx==0.28.1
PyMySQL==1.1.2
aiomysql==0.3.2
cffi==2.0.0
cryptography==46.0.3
pycparser==2.23
async-timeout==5.0.1
redis==7.1.0
dbutils==3.1.2
celery==5.6.1
amqp==5.3.1
billiard==4.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
kombu==5.6.2
packaging==25.0
prompt-toolkit==3.0.52
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
charset_normalizer==3.4.4
requests==2.32.5
urllib3==2.6.2
pika==1.3.2
MarkupSafe==3.0.3
jinja2==3.1.6